+++ src/components/pages/CompanyData/DetailsOverlay.tsx
