+++ src/components/pages/CompanyData/StatusInformation.tsx
