+++ src/components/pages/CompanySubscriptions/CompanySubscriptionDetail.tsx
@@ -34,11 +34,11 @@ import {
 import CommonService from 'services/CommonService'
 import { useState, useEffect } from 'react'
 import { getApiBase } from 'services/EnvironmentService'
-import CompanySubscriptionTechnical from './components/CompanySubscriptionTechnical'
-import CompanySubscriptionHeader from './components/CompanySubscriptionHeader'
-import CompanySubscriptionContent from './components/CompanySubscriptionContent'
-import CompanySubscriptionPrivacy from './components/CompanySubscriptionPrivacyContent'
-import CompanySubscriptionDocument from './components/CompanySubscriptionDocument'
+import CompanySubscriptionTechnical from './CompanySubscriptionTechnical'
+import CompanySubscriptionHeader from './CompanySubscriptionHeader'
+import CompanySubscriptionContent from './CompanySubscriptionContent'
+import CompanySubscriptionPrivacy from './CompanySubscriptionPrivacyContent'
+import CompanySubscriptionDocument from './CompanySubscriptionDocument'
 
 export default function CompanySubscriptionDetail() {
   const navigate = useNavigate()
