+++ src/components/pages/CompanySubscriptions/CompanySubscriptionPrivacyContent/style.scss
