+++ src/components/pages/PartnerNetwork/BusinessPartnerDetailOverlay/BusinessPartnerDetail.tsx
