+++ src/components/pages/PartnerNetwork/BusinessPartnerDetailOverlay/BusinessPartnerDetailContent.tsx
