+++ src/components/pages/PartnerNetwork/BusinessPartnerDetailOverlay/DetailGridRow.tsx
