+++ src/components/pages/PartnerNetwork/BusinessPartnerDetailOverlay/GridTitle.tsx
