+++ src/components/pages/ServiceMarketplaceDetail/MarketplaceHeader/index.tsx
@@ -25,7 +25,7 @@ import { useParams } from 'react-router-dom'
 import type { ServiceRequest } from 'features/serviceMarketplace/serviceApiSlice'
 import { show } from 'features/control/overlay'
 import { OVERLAYS, ROLES } from 'types/Constants'
-import './MarketplaceHeader.scss'
+import './style.scss'
 import { setSuccessType } from 'features/serviceMarketplace/slice'
 import { getAssetBase } from 'services/EnvironmentService'
 import { Box } from '@mui/material'
