+++ src/components/pages/ServiceMarketplaceDetail/MarketplaceSubscription/index.tsx
@@ -21,7 +21,7 @@
 import { useTranslation } from 'react-i18next'
 import { Chip, Typography } from '@catena-x/portal-shared-components'
 import type { ServiceRequest } from 'features/serviceMarketplace/serviceApiSlice'
-import './MarketplaceSubscription.scss'
+import './style.scss'
 
 export default function MarketplaceSubscription({
   item,
